expanded maximum data collection to 16,7M and created initail gui

+++ Radar_controller_example/Example_Any_Frequency/Drivers/AD7676/Inc/ad7676.h
@@ -38,9 +38,9 @@ void ad7676_read_samples(uint32_t samples);
 
 void ad7676_read_continuous(bool enable);
 
-void ad7676_display_samples(uint16_t awaited_samples, uint16_t* received_samples, void (*displayFunction)(char* message));
+void ad7676_display_samples(uint32_t awaited_samples, uint16_t* received_samples, void (*displayFunction)(char* message));
 
-void ad7676_send_samples(uint16_t awaited_samples, uint16_t* received_samples, UART_HandleTypeDef* huart2);
+void ad7676_send_samples(uint32_t awaited_samples, uint16_t* received_samples, UART_HandleTypeDef* huart2);
 
 void ad7676_read_one_sample();
